Adds after insert/update to trigger events.
Adds same entry methods in the handler, and logic to call on new createEventRegistrationFromBatch() method to insert corresponding causeview__Event_Registration__c records when the causeview__Gift_Type__c is "Event Registration" and the causeview__GiftBatch__c value is not null (indicating that the transaction gift record was created from a batch process).

Updates test class with methods for insert and update to validate respective functionalty.