Fixes Candidate Signature
Since we have no way of reliably displaying SVG or JSON Lines as images
in an email, these changes allow us to convert a stored Base64 string
(sent from Magentrix) as a BLOB (png) which can then be used as an
attachment and inline image in the VF template.

The component allows us to dynamically select the relevant image (a
Document) via the AttachmentUtility class.

The trigger and handler on the Candidate object perform the conversion
of the base64 string and storage of the resulting png as a Document in
the “Applicant Signatures” folder.  The image uses the Candidate Record
Name (not the actual applicant’s name) as the Document name, so that we
can use that unique value as key in various maps and queries.